requirements.txt
streamlitpandasTabulateseabornmatplotlibscikit-learngoogle-generativeailangchainlangchain-experimentallangchain-google-genaiopenpyxl